# GH-98686: Get rid of "adaptive" and "quick" instructions