--- a/Lib/opcode.py
+++ b/Lib/opcode.py
@@ -278,7 +278,6 @@ _nb_ops = [
 
 _specializations = {
     "BINARY_OP": [
-        "BINARY_OP_ADAPTIVE",
         "BINARY_OP_ADD_FLOAT",
         "BINARY_OP_ADD_INT",
         "BINARY_OP_ADD_UNICODE",
@@ -289,14 +288,12 @@ _specializations = {
         "BINARY_OP_SUBTRACT_INT",
     ],
     "BINARY_SUBSCR": [
-        "BINARY_SUBSCR_ADAPTIVE",
         "BINARY_SUBSCR_DICT",
         "BINARY_SUBSCR_GETITEM",
         "BINARY_SUBSCR_LIST_INT",
         "BINARY_SUBSCR_TUPLE_INT",
     ],
     "CALL": [
-        "CALL_ADAPTIVE",
         "CALL_PY_EXACT_ARGS",
         "CALL_PY_WITH_DEFAULTS",
         "CALL_BOUND_METHOD_EXACT_ARGS",
@@ -316,7 +313,6 @@ _specializations = {
         "CALL_NO_KW_TYPE_1",
     ],
     "COMPARE_OP": [
-        "COMPARE_OP_ADAPTIVE",
         "COMPARE_OP_FLOAT_JUMP",
         "COMPARE_OP_INT_JUMP",
         "COMPARE_OP_STR_JUMP",
@@ -325,12 +321,10 @@ _specializations = {
         "EXTENDED_ARG_QUICK",
     ],
     "FOR_ITER": [
-        "FOR_ITER_ADAPTIVE",
         "FOR_ITER_LIST",
         "FOR_ITER_RANGE",
     ],
     "LOAD_ATTR": [
-        "LOAD_ATTR_ADAPTIVE",
         # These potentially push [NULL, bound method] onto the stack.
         "LOAD_ATTR_CLASS",
         "LOAD_ATTR_GETATTRIBUTE_OVERRIDDEN",
@@ -353,12 +347,10 @@ _specializations = {
         "LOAD_FAST__LOAD_FAST",
     ],
     "LOAD_GLOBAL": [
-        "LOAD_GLOBAL_ADAPTIVE",
         "LOAD_GLOBAL_BUILTIN",
         "LOAD_GLOBAL_MODULE",
     ],
     "STORE_ATTR": [
-        "STORE_ATTR_ADAPTIVE",
         "STORE_ATTR_INSTANCE_VALUE",
         "STORE_ATTR_SLOT",
         "STORE_ATTR_WITH_HINT",
@@ -368,12 +360,10 @@ _specializations = {
         "STORE_FAST__STORE_FAST",
     ],
     "STORE_SUBSCR": [
-        "STORE_SUBSCR_ADAPTIVE",
         "STORE_SUBSCR_DICT",
         "STORE_SUBSCR_LIST_INT",
     ],
     "UNPACK_SEQUENCE": [
-        "UNPACK_SEQUENCE_ADAPTIVE",
         "UNPACK_SEQUENCE_LIST",
         "UNPACK_SEQUENCE_TUPLE",
         "UNPACK_SEQUENCE_TWO_TUPLE",