--- a/Lib/opcode.py
+++ b/Lib/opcode.py
@@ -317,9 +317,6 @@ _specializations = {
         "COMPARE_OP_INT_JUMP",
         "COMPARE_OP_STR_JUMP",
     ],
-    "EXTENDED_ARG": [
-        "EXTENDED_ARG_QUICK",
-    ],
     "FOR_ITER": [
         "FOR_ITER_LIST",
         "FOR_ITER_RANGE",