--- a/Lib/opcode.py
+++ b/Lib/opcode.py
@@ -319,6 +319,9 @@ _specializations = {
         "COMPARE_OP_INT_JUMP",
         "COMPARE_OP_STR_JUMP",
     ],
+    "EXTENDED_ARG": [
+        "EXTENDED_ARG_QUICK",
+    ],
     "FOR_ITER": [
         "FOR_ITER_LIST",
         "FOR_ITER_RANGE",