… Include/internal/pycore_opcode.h
… Include/opcode.h
… Lib/test/test__opcode.py
… Lib/test/test_dis.py
… Lib/test/test_embed.py
… Misc/NEWS.d/next/Core and Builtins/2022-11-06-00-47-11.gh-issue-98686.DBDy6U.rst
… Python/adaptive.md
… Python/compile.c
… Python/generated_cases.c.h
… Python/opcode_targets.h
… Python/specialize.c
… Tools/c-analyzer/cpython/ignored.tsv
… Tools/cases_generator/generate_cases.py