This gets us one step closer to skipping the quickening step entirely for new code objects... with this change, quickening only involves inserting superinstructions and initializing warmup counters. We do this by getting rid of the `EXTENDED_ARG_QUICK` instruction and making all specializable opcodes contain their own adaptive logic.

Getting this right is a bit tricky, since there are four cases where we want to execute an unquickened instruction:
- When the instruction is warming up.
- When a specialized instruction fails a guard.
- When the instruction is backing off after a failed specialization attempt.
- When we're tracing.

The key insight here is that the logic is identical for the first three cases:
- Check if the counter is zero.
  - If so, try to specialize.
  - If not, decrement the counter and run the instruction.

All that we need to do is change the miss counters for specialized instructions to use the same format as the adaptive backoff counter, and the same code paths can be shared. We skip all of this in the fourth case (tracing) with a simple `if (!cframe.use_tracing) { ... }` guard around the adaptive code (maybe there's a clever way of avoiding this branch, but I doubt it's actually very expensive in practice).

Finally, as an added bonus, merging these code paths allows specialization misses to jump directly into the unquickened instructions, rather than using an indirect jump through a shared `miss` block.

<!-- gh-issue-number: gh-98686 -->
* Issue: gh-98686
<!-- /gh-issue-number -->